requirements.txt
openai>=1.30.0
google-generativeai>=0.7.0
python-dotenv>=1.0.0
fastapi>=0.110.0
uvicorn[standard]>=0.29.0
python-multipart>=0.0.9
jinja2>=3.1.0
